tweaked data cleaning in build file, added tests for output of build file;
