+++ db-checkbook/build_scripts/build.py
@@ -1,0 +1,191 @@
+import pandas as pd
+import geopandas as gpd
+import re
+from pathlib import Path
+from sqlalchemy import create_engine, text
+
+_curr_file_path = Path(__file__).resolve()
+LIB_DIR = _curr_file_path.parent.parent / '.library'
+SQL_QUERY_DIR = _curr_file_path.parent.parent / 'sql_query'
+
+DB_URL = 'sqlite:///checkbook.db'
+ENGINE = create_engine(DB_URL)
+
+def _merge_cpdb_geoms() -> gpd.GeoDataFrame: 
+    """
+    :return: mergedf cpdb geometries
+    """
+    def extract_year(filename):
+        # TODO: error checking
+        match = re.search(r'^\d{4}', filename)
+        if match:
+            return int(match.group())
+        return None
+    
+    subdir_list = [p.name for p in LIB_DIR.iterdir() if p.is_dir()]
+    subdir_list = sorted(subdir_list, key=lambda x: extract_year(x), reverse=True) # sort by year
+    gdf_list = []
+    
+    for f in subdir_list:
+        gdf = gpd.read_file(LIB_DIR / f)
+        gdf_list.append(gdf)
+
+    all_cpdb_geoms = pd.concat(gdf_list)
+    # NOTE: keeping the latest geometry when there are multiple
+    all_cpdb_geoms.drop_duplicates(subset='maprojid', keep='first', inplace=True, ignore_index=True)
+    return all_cpdb_geoms
+
+def _clean_checkbook() -> pd.DataFrame:
+    """
+    :return: cleaned checkbook nyc data
+    """
+    df = pd.read_csv(LIB_DIR / 'nycoc_checkbook.csv')
+    df.columns = df.columns.str.replace(' ', '_')
+    df.columns = df.columns.str.lower()
+    # NOTE: This data cleaning is NOT complete, and we should investigate other cases where we should omit data
+    df = df[df['check_amount']<99000000]
+    df = df[df['check_amount']>=0]
+    # taking out white space from capital project for join with cpdb 
+    """
+    remove last three digits and any trailing whitespace from `Capital Project`
+    Example: `Capital Project` = '998CAP2024  005' -> `FMS ID` = '998CAP2024'
+    \s*: matches zero or more whitespace characters
+    d+: matches one or more digits
+    $: assert position at the end of the string
+    """
+    df['fms_id'] = df['capital_project'].str.replace(r'\s*\d+$','') # QA this output because seems this causes an issue with SCA data
+    return df
+
+def _group_checkbook() -> pd.DataFrame: 
+    """
+    :return: checkbook nyc data grouped by capital project
+    """
+    data = _clean_checkbook()
+
+    def fn_join_vals(x):
+        return ';'.join([y for y in list(x) if pd.notna(y)])
+
+    cols_for_grouping = ['fms_id']
+    cols_for_limiting = cols_for_grouping + [
+        'contract_purpose', 
+        'agency', 
+        'budget_code', 
+        'check_amount'
+    ]
+    df_limited_cols = data.loc[:, cols_for_limiting]
+    agg_dict = {
+        'check_amount': 'sum',
+        'contract_purpose': fn_join_vals,
+        'budget_code': fn_join_vals,
+        'agency': fn_join_vals
+    }
+    df = df_limited_cols.groupby(cols_for_grouping, as_index=False).agg(agg_dict)
+    return df
+
+def _join_checkbook_geoms(df: pd.DataFrame, cpdb_geoms: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
+    """
+    :param df: Checkbook NYC data collapsed on FMS ID
+    :param cpdb_geoms: final versions of archived CPDB geometries from every year, and the most recent geometry for the current year
+    :return: CPDB geometries left-joined onto Checkbook NYC data 
+    """
+    merged = df.merge(cpdb_geoms, how='left', left_on='fms_id', right_on='maprojid', indicator=True)
+    gdf = gpd.GeoDataFrame(merged, geometry='geometry')
+    return gdf
+
+def _assign_checkbook_category(df: pd.DataFrame) -> pd.DataFrame:
+    """
+    param df: cleaned and collapsed checkbook NYC data 
+    return: pandas df of checkbook data with category assignment based on specified col 
+    """
+    queries = [SQL_QUERY_DIR / 'query_itt_vehicles_equipment.sql', SQL_QUERY_DIR / 'query_lump_sum.sql', SQL_QUERY_DIR / 'query_fixed_asset.sql']
+    target_cols = {'budget_code': 'bc_category', 'contract_purpose': 'cp_category'}
+    df['bc_category'] = None
+    df['cp_category'] = None
+    
+    with ENGINE.connect() as conn:
+        df.to_sql('capital_projects', ENGINE, if_exists='replace', index=False)
+        for k, v in target_cols.items():
+            with open(SQL_QUERY_DIR / 'query.sql', 'r') as query_file:
+                query = query_file.read()
+            query = query.replace('COLUMN', k)
+            query = query.replace('col_category', v)
+            queries = [q.strip() for q in query.split(';') if q.strip()]
+            for q in queries:
+                conn.execute(text(q))
+
+        ret = pd.read_sql_table('capital_projects', conn)
+
+    return ret
+
+def _clean_joined_checkbook_cpdb(gdf: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
+    """
+    :param gdf: joined cpdb and checkbook nyc data
+    :return: cleaned joined checkbook cpdb data 
+    """
+    gdf.rename(columns={'typecatego':'cpdb_category'}, inplace=True)
+    gdf['has_geometry'] = gdf['_merge'].map(lambda x: x == 'both')
+    gdf['cpdb_category'].fillna('None', inplace=True)
+    gdf['bc_category'].fillna('None', inplace=True)
+    gdf['cp_category'].fillna('None', inplace=True)
+    gdf.drop('_merge', axis=1, inplace=True)
+    gdf = _limit_cols(gdf)
+    return gdf
+
+def _limit_cols(gdf: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
+    """
+    :param gdf: joined checkbook nyc and cpdb data
+    :return: geopandas df of joined checkbook cpdb data with cols limited/reordered
+    """
+    return gdf[[
+        'fms_id', 
+        'check_amount', 
+        'contract_purpose', 
+        'budget_code', 
+        'agency', 
+        'bc_category', 
+        'cp_category', 
+        'cpdb_category', 
+        'ccpversion',
+        'maprojid',
+        'magency',
+        'magencyacr',
+        'projectid',
+        'descriptio',
+        'geomsource',
+        'dataname',
+        'datasource',
+        'datadate',
+        'geometry',
+        'cartodb_id',
+        'has_geometry'
+    ]]
+
+def _assign_final_category(gdf: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
+    """
+    return: geopandas gdf with merged checkbook cpdb data and 
+    final category assignment using high sensitivity fixed asset method
+    """ 
+    cols = ['cpdb_category', 'bc_category', 'cp_category']
+    cats = ['Fixed Asset', 'ITT, Vehicles and Equipment', 'Lump Sum', 'None']
+
+    gdf['final_category'] = gdf[cols].apply(lambda row: cats[next((i for i, cat in enumerate(cats) if cat in row.values), 3)], axis=1)
+    return gdf
+
+def run_build() -> pd.DataFrame:
+    """
+    :return: historical spending data
+    """
+    cpdb_geoms = _merge_cpdb_geoms()
+    grouped_checkbook = _group_checkbook()
+    cat_checkbook = _assign_checkbook_category(grouped_checkbook)
+    joined_data = _join_checkbook_geoms(cat_checkbook, cpdb_geoms)
+    cleaned_data = _clean_joined_checkbook_cpdb(joined_data)
+    final_data = _assign_final_category(cleaned_data)
+    return final_data
+
+if __name__ == "__main__":
+    print("started build ...")
+    final_data = run_build()
+    print('.. done!')
+    fp = _curr_file_path.parent.parent / 'output' / 'temp_historical_spend.csv'
+    final_data.to_csv(fp)
